Moving the ReducedDocumentText class to jcore-utilities.
Also renaming the class to JCoReCutAwayText.
