+++ jcore-utilities/src/main/java/de/julielab/jcore/utility/JCoReCutAwayDocumentText.java
@@ -1,4 +1,4 @@
-package de.julielab.jcore.ae.jsbd.main;
+package de.julielab.jcore.utility;
 
 import java.util.NavigableMap;
 import java.util.Set;
@@ -7,36 +7,48 @@ import java.util.Map.Entry;
 
 import org.apache.uima.jcas.JCas;
 
-import de.julielab.jcore.utility.JCoReAnnotationIndexMerger;
-
-public class ReducedDocumentText {
+/**
+ * This class is helpful when some parts of the CAS document text should be cut
+ * out according to a set of specific annotations. The class then represents the
+ * document text that results from cutting out said text passages. It offers a
+ * method to return the actual text string and a method to map the character
+ * offsets of the compacted string to the original CAS document text.
+ * 
+ * @author faessler
+ *
+ */
+public class JCoReCutAwayDocumentText {
 	private NavigableMap<Integer, Integer> spanSumMap;
-	private String reducedText;
+	private String cutText;
 
-	public ReducedDocumentText(JCas cas, Set<String> ignoredTypes) throws ClassNotFoundException {
-		buildMap(cas, ignoredTypes);
+	public JCoReCutAwayDocumentText(JCas cas, Set<String> cutAwayTypes) throws ClassNotFoundException {
+		buildMap(cas, cutAwayTypes);
 	}
 
 	/**
-	 * Creates a map that maps those positions of the reduced text that correspond
-	 * to an intermediate next position after an ignored annotation in the original
-	 * text to the sum of ranges covered by ignored annotations up to the original
+	 * Creates a map that maps those positions of the small-cut text that correspond
+	 * to an intermediate next position after a cut-away annotation in the original
+	 * text to the sum of ranges covered by cut-away annotations up to the original
 	 * offset.
 	 * 
 	 * @param cas
-	 * @param ignoredTypes
+	 *            The CAS for create a cut-away document text for.
+	 * @param cutAwayTypes
+	 *            The qualified type names of the annotations whose covered text
+	 *            should be cut away.
 	 * @throws ClassNotFoundException
+	 *             If <tt>cutAwayTypes</tt> contains type identifiers to non-existing types.
 	 */
-	public void buildMap(JCas cas, Set<String> ignoredTypes) throws ClassNotFoundException {
+	public void buildMap(JCas cas, Set<String> cutAwayTypes) throws ClassNotFoundException {
 		StringBuilder sb = new StringBuilder();
 		spanSumMap = new TreeMap<>();
 		spanSumMap.put(0, 0);
-		JCoReAnnotationIndexMerger merger = new JCoReAnnotationIndexMerger(ignoredTypes, true, null, cas);
+		JCoReAnnotationIndexMerger merger = new JCoReAnnotationIndexMerger(cutAwayTypes, true, null, cas);
 		int spanSum = 0;
 		int lastBegin = 0;
 		int lastEnd = -1;
 		// For each ignored annotation, there could be following annotations overlapping
-		// with the first, effectively enlengthening the ignored span. Thus, we iterate
+		// with the first, effectively enlargeing the ignored span. Thus, we iterate
 		// until we find an ignored annotation the has a positive (not 0) distance to a
 		// previous one. Then, we store the length of the span of ignored annotations
 		// for the largest end of the previous annotations.
@@ -64,15 +76,15 @@ public class ReducedDocumentText {
 		}
 		if (lastEnd < cas.getDocumentText().length())
 			sb.append(cas.getDocumentText().substring(lastEnd, cas.getDocumentText().length()));
-		reducedText = sb.toString();
+		cutText = sb.toString();
 	}
 
-	public int getOriginalOffsetForReducedOffset(int reducedOffset) {
-		Entry<Integer, Integer> floorEntry = spanSumMap.floorEntry(reducedOffset);
-		return reducedOffset + floorEntry.getValue();
+	public int getOriginalOffsetForCutAwayOffset(int cutAwayOffset) {
+		Entry<Integer, Integer> floorEntry = spanSumMap.floorEntry(cutAwayOffset);
+		return cutAwayOffset + floorEntry.getValue();
 	}
 
 	public String getReducedText() {
-		return reducedText;
+		return cutText;
 	}
 }
